zuj-name-app/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-Levenshtein